// lib/types.ts
export interface User {
    _id: string
    names: string
    email: string
    gender: string
    role: string
    class: string
}

export interface Report {
    _id: string
    absent: string
    away: string
    created: string
    helped: string
    presents: string
    sabbath_week: string
    sick: string
    startedSabbath: string
    studied7times: string
    visited: string
    vistors: string
    wereHelped: string
    wereVisted: string
    year: string
}